requirements.txt
databricks-sdk>=0.25.0
python-dotenv>=1.0.0
typer[all]>=0.12.0
rich>=13.7.0
pytest>=8.0.0
pytest-mock>=3.12.0
